requirements.txt
fastapi
requests
dune-client
joblib
uvicorn
redis

python-dotenv